Fixed 64-bit compiler errors.
Discovered that the HG message definition did not match.
This makes me want to get to one of two solutions that will resolve this
issue:
1) Change the MACRO definition table to generate the typelist
internally. This would eliminate the need for a separately defined
typelist. The Bit set automatically generates its own typelist
(integer_sequence template).
2) Add warning tests to the HG_ MACROS that report if a different type
is supplied in the MACRO than was defined in the Typelist. The only
reason the types are listed in the MACRO is to provide a source of
reference for the user that matches the name with the type.

+++ U232/common/msgtest_typelist.h
@@ -58,7 +58,7 @@ typedef TypeList
   uint32_t,
   uint8_t,
   nested_format_t,
-  std::array<size_t, 3>
+  std::array<uint32_t, 3>
 > base_format_t;
